feat : BigPoster.jsx 너무 커서 컴포넌트 분리
