+++ src/utils/preTicketingPeriod.js
@@ -1,7 +1,6 @@
 // NOTYET(사전예매 시작 전) , ING(사전예매 기간 중), END(사전예매 종료)
 export default function PreTicketingPeriod(_time) {
   let PreTPeriodState = '';
-  console.log('_time : ', _time.length);
   if (_time.length !== 0) {
     if (Date.now() >= _time[0] && Date.now() < _time[1]) {
       PreTPeriodState = '중';
